Intial work on 1.9

+++ java/squeek/applecore/asm/reference/BlockCactusModifications.java
@@ -2,7 +2,7 @@ package squeek.applecore.asm.reference;
 
 import net.minecraft.block.BlockCactus;
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.util.BlockPos;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.eventhandler.Event;
 import squeek.applecore.asm.Hooks;
@@ -26,18 +26,18 @@ public class BlockCactusModifications extends BlockCactus
 			// added && ...
 			if (i < 3 && Hooks.fireAllowPlantGrowthEvent(this, world, pos, state, rand) != Event.Result.DENY)
 			{
-				int j = (state.getValue(AGE)).intValue();
+				int j = state.getValue(AGE);
 
 				if (j == 15)
 				{
 					world.setBlockState(pos.up(), this.getDefaultState());
-					IBlockState iblockstate = state.withProperty(AGE, Integer.valueOf(0));
+					IBlockState iblockstate = state.withProperty(AGE, 0);
 					world.setBlockState(pos, iblockstate, 4);
 					this.onNeighborBlockChange(world, pos.up(), iblockstate, this);
 				}
 				else
 				{
-					world.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(j + 1)), 4);
+					world.setBlockState(pos, state.withProperty(AGE, j + 1), 4);
 				}
 
 				// added line
