+++ java/squeek/applecore/asm/reference/BlockNetherWartModifications.java
@@ -2,7 +2,7 @@ package squeek.applecore.asm.reference;
 
 import net.minecraft.block.BlockNetherWart;
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.util.BlockPos;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.eventhandler.Event.Result;
 import squeek.applecore.asm.Hooks;
@@ -19,12 +19,12 @@ public class BlockNetherWartModifications extends BlockNetherWart
 		// added line
 		Result allowGrowthResult = Hooks.fireAllowPlantGrowthEvent(this, world, pos, state, rand);
 
-		int i = (state.getValue(AGE)).intValue();
+		int i = state.getValue(AGE);
 		
 		// changed if
 		if (i < 3 && (allowGrowthResult == Result.ALLOW || (allowGrowthResult == Result.DEFAULT && rand.nextInt(10) == 0)))
 		{
-			state = state.withProperty(AGE, Integer.valueOf(i + 1));
+			state = state.withProperty(AGE, i + 1);
 			world.setBlockState(pos, state, 2);
 
 			// added line
